# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "MetaWall API",
    "description": "MetaWall 元宇宙社交圈 API 文件"
  },
  "host": "https://fierce-headland-11812.herokuapp.com",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "scheme": "bearer",
      "bearerFormat": "JWT",
      "in": "headers",
      "name": "authorization",
      "description": "請加上 API Token"
    },
    "api_key": {
      "type": "apiKey",
      "in": "header",
      "name": "X-API-KEY",
      "description": "system administrators's access permission"
    }
  },
  "consumes": [],
  "produces": [],
  "paths": {
    "/check/check": {
      "post": {
        "tags": [
          "Token 驗證"
        ],
        "description": "驗證 token 正確與否 API",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/users/": {
      "get": {
        "tags": [
          "後台：Users - 用戶"
        ],
        "description": "取得所有用戶資訊 API",
        "parameters": [],
        "responses": {
          "200": {
            "description": "管理者成功取得所有使用者資料",
            "schema": {
              "$ref": "#/definitions/getAllUsers_Schema"
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "delete": {
        "tags": [
          "後台：Users - 用戶"
        ],
        "description": "刪除所有用戶資訊 API",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/posts/": {
      "get": {
        "tags": [
          "Posts - 貼文"
        ],
        "description": "取得所有貼文 API",
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功取得所有貼文內容",
            "schema": {
              "$ref": "#/definitions/getPosts_Schema"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Posts - 貼文"
        ],
        "description": "新增貼文 API",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "type": "object",
            "required": true,
            "description": "資料格式",
            "schema": {
              "$ref": "#/definitions/createdPosts_Schema"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功新增一篇貼文",
            "schema": {
              "$ref": "#/definitions/getPosts_Schema"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "後台：Posts - 貼文"
        ],
        "description": "刪除所有貼文 API",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/comments/": {
      "post": {
        "tags": [
          "Comments - 貼文評論"
        ],
        "description": "新增單筆貼文評論資訊 API",
        "parameters": [],
        "responses": {}
      },
      "delete": {
        "tags": [
          "後台：Comments - 貼文評論"
        ],
        "description": "刪除所有貼文評論資訊 API",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/upload/": {
      "post": {
        "tags": [
          "Other - 上傳圖片"
        ],
        "description": "上傳圖片到 imgUr 圖床 API",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "form-data 資料格式",
            "type": "string"
          }
        ],
        "responses": {}
      }
    }
  },
  "definitions": {
...
    "getAllUsers_Schema": {
      "type": "object",
      "properties": {
        "status": {
          "type": "boolean",
          "example": true
        },
        "data": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "role": {
                "type": "string",
                "example": "user2"
              },
              "_id": {
                "type": "string",
                "example": "11b1561v5d02zz2bfr0d1132"
              },
              "name": {
                "type": "string",
                "example": "NewUser"
              },
              "avatar": {
                "type": "string",
                "example": "https://i.imgur.com/K3dyy79.png"
              },
              "createdAt": {
                "type": "string",
                "example": "2022-06-02T13:30:27.204Z"
              },
              "updatedAt": {
                "type": "string",
                "example": "2022-06-02T13:30:27.204Z"
              }
            }
          }
        }
      }
    },
    "getPosts_Schema": {
      "type": "object",
      "properties": {
        "status": {
          "type": "boolean",
          "example": true
        },
        "data": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string",
                "example": "6298cd452f75e1bf5c347d55"
              },
              "author": {
                "type": "object",
                "properties": {
                  "_id": {
                    "type": "string",
                    "example": "6298bdcc86d7d2a709c289e1"
                  },
                  "name": {
                    "type": "string",
                    "example": "波吉"
                  },
                  "avatar": {
                    "type": "string",
                    "example": "https://www.nj.com/resizer/iqV2J-QFgh0227ybHBor4exTVBk=/800x0/smart/cloudfront-us-east-1.images.arcpublishing.com/advancelocal/SJGKVE5UNVESVCW7BBOHKQCZVE.jpg"
                  }
                }
              },
              "content": {
                "type": "string",
                "example": "我一定要成為很棒棒的國王！"
              },
              "tags": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "image": {
                "type": "string",
                "example": "https://images.unsplash.com/photo-1499026008573-50eedca8407b?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=761&q=80"
              },
              "likes": {
                "type": "number",
                "example": 0
              },
              "comments": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "example": "6299d7788e41247531c5b899"
                    },
                    "articleID": {
                      "type": "string",
                      "example": "6298cd452f75e1bf5c347d55"
                    },
                    "author": {
                      "type": "string",
                      "example": "6298bb7386d7d2a709c289de"
                    },
                    "body": {
                      "type": "string",
                      "example": "加油～你可以的！"
                    },
                    "createdAt": {
                      "type": "string",
                      "example": "2022-06-03T09:42:16.882Z"
                    },
                    "updatedAt": {
                      "type": "string",
                      "example": "2022-06-03T09:42:16.882Z"
                    }
                  }
                }
              },
              "privacy": {
                "type": "string",
                "example": "private"
              },
              "createdAt": {
                "type": "string",
                "example": "2022-06-02T14:46:29.020Z"
              },
              "updatedAt": {
                "type": "string",
                "example": "2022-06-03T10:16:29.092Z"
              }
            }
          }
        }
      }
    },
    "createdPosts_Schema": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string",
          "example": "我一定要成為很棒棒的國王！"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "image": {
          "type": "string",
          "example": "https://images.unsplash.com/photo-1499026008573-50eedca8407b?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=761&q=80"
        },
        "privacy": {
          "type": "string",
          "example": "private"
        }
      },
      "required": [
        "content"
      ]
    },
...
  }
}